feat: implement shader parsing utilities and module registry
